Add generate all convenience method
