+++ crates/ubrn_cli/src/ios.rs
@@ -236,4 +236,8 @@ impl IOsArgs {
     pub(crate) fn project_config(&self) -> Result<ProjectConfig> {
         self.config.clone().try_into()
     }
+
+    pub(crate) fn config(&self) -> Utf8PathBuf {
+        self.config.clone()
+    }
 }
